# ReadOnlySequence: Don't leak the flags that are on _startInteger and _endInteger.