--- a/src/libraries/System.Memory/src/System/Buffers/ReadOnlySequence.cs
+++ b/src/libraries/System.Memory/src/System/Buffers/ReadOnlySequence.cs
@@ -62,7 +62,7 @@ namespace System.Buffers
         public SequencePosition Start
         {
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            get => new SequencePosition(_startObject, _startInteger);
+            get => new SequencePosition(_startObject, _startInteger & ReadOnlySequence.IndexBitMask);
         }
 
         /// <summary>
@@ -71,7 +71,7 @@ namespace System.Buffers
         public SequencePosition End
         {
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            get => new SequencePosition(_endObject, _endInteger);
+            get => new SequencePosition(_endObject, _endInteger & ReadOnlySequence.IndexBitMask);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
@@ -660,6 +660,10 @@ namespace System.Buffers
 
     internal static class ReadOnlySequence
     {
+        /// <summary>
+        /// Flag that allow to encode the <see cref="ReadOnlySequence{T}.SequenceType"/>.
+        /// </summary>
+        /// <seealso cref="ReadOnlySequence{T}.GetSequenceType"/>
         public const int FlagBitMask = 1 << 31;
         public const int IndexBitMask = ~FlagBitMask;