--- a/src/libraries/System.Memory/tests/ReadOnlyBuffer/ReadOnlySequenceTests.byte.cs
+++ b/src/libraries/System.Memory/tests/ReadOnlyBuffer/ReadOnlySequenceTests.byte.cs
@@ -187,6 +187,29 @@ namespace System.Memory.Tests
             Assert.Equal(buffer.End, buffer.GetPosition(5));
         }
 
+        [Fact]
+        public void Start_EqualToGetPosition0()
+        {
+            ReadOnlySequence<byte> buffer = Factory.CreateOfSize(5);
+            Assert.Equal(buffer.Start, buffer.GetPosition(0));
+        }
+
+        [Fact]
+        public void InnerPositionAreNotEqualToEnd()
+        {
+            ReadOnlySequence<byte> buffer = Factory.CreateOfSize(3);
+            Assert.NotEqual(buffer.GetPosition(1), buffer.End);
+            Assert.NotEqual(buffer.GetPosition(2), buffer.End);
+        }
+
+        [Fact]
+        public void InnerPositionAreNotEqualToStart()
+        {
+            ReadOnlySequence<byte> buffer = Factory.CreateOfSize(3);
+            Assert.NotEqual(buffer.GetPosition(1), buffer.Start);
+            Assert.NotEqual(buffer.GetPosition(2), buffer.Start);
+        }
+
         public static TheoryData<Func<ReadOnlySequence<byte>, ReadOnlySequence<byte>>> ValidSliceCases => new TheoryData<Func<ReadOnlySequence<byte>, ReadOnlySequence<byte>>>
         {
             b => b.Slice(5),