Fixes #42999.
Currently the ReadOnlySequence.End has sometimes a flag that causes odd behavior like:  
`GetPosition(size) != End`
`ReadOnlySequence<byte>.Empty.Start != ReadOnlySequence<byte>.Empty.End`  
This causes more odd behavior later on types relying on the ROSequence.

My fix simply removes the flags when we return a SequencePosition in .Start and .End.