@Kuinox 2021-02-08
And here the benchmarks:  
```
master:
|                          Method |      Mean |     Error |    StdDev |    Median |       Min |       Max 
|-------------------------------- |----------:|----------:|----------:|----------:|----------:|----------:
|              IterateTryGetArray |  6.292 ns | 0.0532 ns | 0.0471 ns |  6.273 ns |  6.255 ns |  6.408 ns 
|             IterateForEachArray | 17.566 ns | 0.1753 ns | 0.1640 ns | 17.485 ns | 17.393 ns | 17.747 ns 
|         IterateGetPositionArray | 17.354 ns | 0.1607 ns | 0.1503 ns | 17.349 ns | 17.155 ns | 17.589 ns 
|                      FirstArray |  6.720 ns | 0.1333 ns | 0.2036 ns |  6.737 ns |  6.353 ns |  7.212 ns 
|                  FirstSpanArray |  8.822 ns | 0.1050 ns | 0.0983 ns |  8.786 ns |  8.677 ns |  9.001 ns 
|                 FirstSpanMemory | 11.549 ns | 0.2215 ns | 0.2072 ns | 11.593 ns | 11.202 ns | 11.832 ns 
|          FirstSpanSingleSegment | 18.710 ns | 0.0468 ns | 0.0365 ns | 18.721 ns | 18.661 ns | 18.784 ns 
|            FirstSpanTenSegments | 19.025 ns | 0.1547 ns | 0.1447 ns | 18.955 ns | 18.897 ns | 19.265 ns 
|                      SliceArray |  6.021 ns | 0.0539 ns | 0.0478 ns |  6.034 ns |  5.939 ns |  6.078 ns 
|             IterateTryGetMemory | 24.220 ns | 0.2423 ns | 0.2267 ns | 24.129 ns | 23.900 ns | 24.572 ns 
|            IterateForEachMemory | 33.338 ns | 0.3058 ns | 0.2860 ns | 33.289 ns | 32.997 ns | 33.735 ns 
|        IterateGetPositionMemory | 36.789 ns | 0.4317 ns | 0.4038 ns | 36.501 ns | 36.438 ns | 37.560 ns 
|                     FirstMemory |  7.687 ns | 0.1543 ns | 0.1585 ns |  7.621 ns |  7.490 ns |  7.929 ns 
|                     SliceMemory |  7.946 ns | 0.0651 ns | 0.0609 ns |  7.939 ns |  7.878 ns |  8.045 ns 
|      IterateTryGetSingleSegment | 13.065 ns | 0.1347 ns | 0.1260 ns | 13.051 ns | 12.875 ns | 13.207 ns 
|     IterateForEachSingleSegment | 24.571 ns | 0.2268 ns | 0.2121 ns | 24.497 ns | 24.339 ns | 24.875 ns 
| IterateGetPositionSingleSegment | 23.719 ns | 0.0713 ns | 0.0596 ns | 23.718 ns | 23.598 ns | 23.840 ns 
|              FirstSingleSegment |  4.408 ns | 0.0896 ns | 0.1066 ns |  4.363 ns |  4.305 ns |  4.699 ns 
|              SliceSingleSegment |  6.551 ns | 0.0118 ns | 0.0092 ns |  6.551 ns |  6.533 ns |  6.565 ns 
|        IterateTryGetTenSegments | 39.891 ns | 0.0123 ns | 0.0103 ns | 39.890 ns | 39.875 ns | 39.913 ns 
|       IterateForEachTenSegments | 60.201 ns | 0.3256 ns | 0.3045 ns | 60.111 ns | 59.876 ns | 60.760 ns 
|   IterateGetPositionTenSegments | 77.265 ns | 0.1021 ns | 0.0797 ns | 77.263 ns | 77.107 ns | 77.444 ns 
|                FirstTenSegments |  3.796 ns | 0.0249 ns | 0.0221 ns |  3.801 ns |  3.752 ns |  3.826 ns 
|                SliceTenSegments | 23.177 ns | 0.2675 ns | 0.2234 ns | 23.098 ns | 22.917 ns | 23.662 ns 

PR:  
|                          Method |      Mean |     Error |    StdDev |    Median |       Min |       Max 
|-------------------------------- |----------:|----------:|----------:|----------:|----------:|----------:
|              IterateTryGetArray |  6.242 ns | 0.1057 ns | 0.0937 ns |  6.245 ns |  6.069 ns |  6.368 ns 
|             IterateForEachArray | 17.536 ns | 0.2349 ns | 0.2082 ns | 17.436 ns | 17.378 ns | 18.034 ns 
|         IterateGetPositionArray | 16.901 ns | 0.3560 ns | 0.3957 ns | 16.662 ns | 16.314 ns | 17.492 ns 
|                      FirstArray |  7.023 ns | 0.2879 ns | 0.8444 ns |  6.662 ns |  6.230 ns |  8.954 ns 
|                  FirstSpanArray |  8.850 ns | 0.0681 ns | 0.0637 ns |  8.844 ns |  8.725 ns |  8.942 ns 
|                 FirstSpanMemory | 11.418 ns | 0.2213 ns | 0.2070 ns | 11.402 ns | 10.959 ns | 11.801 ns 
|          FirstSpanSingleSegment | 18.867 ns | 0.1446 ns | 0.1353 ns | 18.790 ns | 18.741 ns | 19.162 ns 
|            FirstSpanTenSegments | 19.079 ns | 0.2027 ns | 0.1896 ns | 18.989 ns | 18.880 ns | 19.387 ns 
|                      SliceArray |  5.985 ns | 0.0697 ns | 0.0652 ns |  5.955 ns |  5.916 ns |  6.134 ns 
|             IterateTryGetMemory | 24.326 ns | 0.5120 ns | 1.3930 ns | 24.035 ns | 22.405 ns | 27.494 ns 
|            IterateForEachMemory | 34.008 ns | 0.7100 ns | 1.0843 ns | 33.843 ns | 32.760 ns | 36.556 ns 
|        IterateGetPositionMemory | 37.690 ns | 0.6607 ns | 0.6180 ns | 38.062 ns | 36.653 ns | 38.565 ns 
|                     FirstMemory |  7.775 ns | 0.1560 ns | 0.2564 ns |  7.630 ns |  7.411 ns |  8.207 ns 
|                     SliceMemory |  8.066 ns | 0.1601 ns | 0.1572 ns |  8.027 ns |  7.872 ns |  8.299 ns 
|      IterateTryGetSingleSegment | 13.360 ns | 0.2759 ns | 0.2581 ns | 13.266 ns | 13.059 ns | 13.731 ns 
|     IterateForEachSingleSegment | 25.066 ns | 0.5198 ns | 0.5105 ns | 25.008 ns | 24.391 ns | 25.715 ns 
| IterateGetPositionSingleSegment | 24.344 ns | 0.2918 ns | 0.2729 ns | 24.542 ns | 23.905 ns | 24.632 ns 
|              FirstSingleSegment |  4.395 ns | 0.0400 ns | 0.0445 ns |  4.389 ns |  4.297 ns |  4.497 ns 
|              SliceSingleSegment |  6.606 ns | 0.0852 ns | 0.0797 ns |  6.594 ns |  6.523 ns |  6.714 ns 
|        IterateTryGetTenSegments | 43.273 ns | 0.8911 ns | 1.4133 ns | 42.584 ns | 41.741 ns | 46.095 ns 
|       IterateForEachTenSegments | 60.702 ns | 0.8259 ns | 0.7726 ns | 60.471 ns | 59.866 ns | 61.716 ns 
|   IterateGetPositionTenSegments | 78.647 ns | 1.2567 ns | 1.1755 ns | 78.457 ns | 77.200 ns | 80.417 ns 
|                FirstTenSegments |  4.037 ns | 0.0831 ns | 0.2100 ns |  3.951 ns |  3.728 ns |  4.398 ns 
|                SliceTenSegments | 23.666 ns | 0.4604 ns | 0.6603 ns | 23.374 ns | 23.042 ns | 25.128 ns 
```